Refactoring network code
