+++ aqt/archives.py
@@ -23,27 +23,8 @@
 import xml.etree.ElementTree as ElementTree
 from logging import getLogger
 
-import requests
-from requests.adapters import HTTPAdapter
-from urllib3 import Retry
-
-from aqt.helper import Settings, altlink
-
-
-class ArchiveConnectionError(Exception):
-    pass
-
-
-class ArchiveListError(Exception):
-    pass
-
-
-class ArchiveDownloadError(Exception):
-    pass
-
-
-class NoPackageFound(Exception):
-    pass
+from aqt.exceptions import ArchiveListError, NoPackageFound
+from aqt.helper import Settings, getUrl
 
 
 class TargetConfig:
@@ -118,8 +99,8 @@ class PackagesList:
                 ext,
             )
             update_xml_url = "{0}{1}Updates.xml".format(self.base, archive_path)
-            r = requests.get(update_xml_url, timeout=self.timeout)
-            self.update_xml = ElementTree.fromstring(r.text)
+            xml_text = getUrl(update_xml_url, self.timeout, self.logger)
+            self.update_xml = ElementTree.fromstring(xml_text)
             for packageupdate in self.update_xml.iter("PackageUpdate"):
                 name = packageupdate.find("Name").text
                 if packageupdate.find("DownloadableArchives").text is not None:
@@ -213,38 +194,8 @@ class QtArchives:
         )
         target_packages.append("qt.{}.{}".format(qt_ver_num, self.arch))
         target_packages.extend(self.mod_list)
-        self._download_update_xml(update_xml_url)
+        self.update_xml_text = getUrl(update_xml_url, self.timeout, self.logger)
         self._parse_update_xml(archive_url, target_packages)
-
-    def _download_update_xml(self, update_xml_url):
-        with requests.Session() as session:
-            retry = Retry(connect=5, backoff_factor=0.5)
-            adapter = HTTPAdapter(max_retries=retry)
-            session.mount("http://", adapter)
-            session.mount("https://", adapter)
-            try:
-                r = requests.get(update_xml_url, allow_redirects=False, timeout=self.timeout)
-                if r.status_code == 302:
-                    newurl = altlink(r.url, r.headers["Location"], logger=self.logger)
-                    self.logger.info("Redirected URL: {}".format(newurl))
-                    r = session.get(newurl, stream=True, timeout=self.timeout)
-            except (
-                ConnectionResetError,
-                requests.exceptions.ConnectionError,
-                requests.exceptions.Timeout,
-            ):
-                raise ArchiveConnectionError()
-            else:
-                if r.status_code == 200:
-                    self.update_xml_text = r.text
-                else:
-                    self.logger.error(
-                        "Download error when access to {}\n"
-                        "Server response code: {}, reason: {}".format(
-                            update_xml_url, r.status_code, r.reason
-                        )
-                    )
-                    raise ArchiveDownloadError("Download error!")
 
     def _parse_update_xml(self, archive_url, target_packages):
         try:
